Removed all data ops in Tensor class. Refactored tests accordingly.
